refactor ZHA
