+++ homeassistant/components/zha/__init__.py
@@ -7,15 +7,15 @@ https://home-assistant.io/components/zha/
 import collections
 import enum
 import logging
-import time
 
 import voluptuous as vol
 
 import homeassistant.helpers.config_validation as cv
 from homeassistant import const as ha_const
-from homeassistant.helpers import discovery, entity
-from homeassistant.util import slugify
+from homeassistant.helpers import discovery
 from homeassistant.helpers.entity_component import EntityComponent
+from homeassistant.components.zha.entities import ZhaDeviceEntity
+from . import const as zha_const
 
 REQUIREMENTS = [
     'bellows==0.7.0',
@@ -145,6 +145,7 @@ class ApplicationListener:
         self._component = EntityComponent(_LOGGER, DOMAIN, hass)
         self._device_registry = collections.defaultdict(list)
         hass.data[DISCOVERY_KEY] = hass.data.get(DISCOVERY_KEY, {})
+        zha_const.populate_data()
 
     def device_joined(self, device):
         """Handle device joined.
@@ -177,8 +178,6 @@ class ApplicationListener:
     async def async_device_initialized(self, device, join):
         """Handle device joined and basic information discovered (async)."""
         import zigpy.profiles
-        import homeassistant.components.zha.const as zha_const
-        zha_const.populate_data()
 
         device_manufacturer = device_model = None
 
@@ -276,7 +275,6 @@ class ApplicationListener:
                                              device_classes, discovery_attr,
                                              is_new_join):
         """Try to set up an entity from a "bare" cluster."""
-        import homeassistant.components.zha.const as zha_const
         if cluster.cluster_id in profile_clusters:
             return
 
@@ -319,227 +317,4 @@ class ApplicationListener:
             DOMAIN,
             {'discovery_key': cluster_key},
             self._config,
-        )
-
-
-class Entity(entity.Entity):
-    """A base class for ZHA entities."""
-
-    _domain = None  # Must be overridden by subclasses
-
-    def __init__(self, endpoint, in_clusters, out_clusters, manufacturer,
-                 model, application_listener, unique_id, **kwargs):
-        """Init ZHA entity."""
-        self._device_state_attributes = {}
-        ieee = endpoint.device.ieee
-        ieeetail = ''.join(['%02x' % (o, ) for o in ieee[-4:]])
-        if manufacturer and model is not None:
-            self.entity_id = "{}.{}_{}_{}_{}{}".format(
-                self._domain,
-                slugify(manufacturer),
-                slugify(model),
-                ieeetail,
-                endpoint.endpoint_id,
-                kwargs.get('entity_suffix', ''),
-            )
-            self._device_state_attributes['friendly_name'] = "{} {}".format(
-                manufacturer,
-                model,
-            )
-        else:
-            self.entity_id = "{}.zha_{}_{}{}".format(
-                self._domain,
-                ieeetail,
-                endpoint.endpoint_id,
-                kwargs.get('entity_suffix', ''),
-            )
-
-        self._endpoint = endpoint
-        self._in_clusters = in_clusters
-        self._out_clusters = out_clusters
-        self._state = None
-        self._unique_id = unique_id
-
-        # Normally the entity itself is the listener. Sub-classes may set this
-        # to a dict of cluster ID -> listener to receive messages for specific
-        # clusters separately
-        self._in_listeners = {}
-        self._out_listeners = {}
-
-        self._initialized = False
-        application_listener.register_entity(ieee, self)
-
-    async def async_added_to_hass(self):
-        """Handle entity addition to hass.
-
-        It is now safe to update the entity state
-        """
-        for cluster_id, cluster in self._in_clusters.items():
-            cluster.add_listener(self._in_listeners.get(cluster_id, self))
-        for cluster_id, cluster in self._out_clusters.items():
-            cluster.add_listener(self._out_listeners.get(cluster_id, self))
-
-        self._initialized = True
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID."""
-        return self._unique_id
-
-    @property
-    def device_state_attributes(self):
-        """Return device specific state attributes."""
-        return self._device_state_attributes
-
-    def attribute_updated(self, attribute, value):
-        """Handle an attribute updated on this cluster."""
-        pass
-
-    def zdo_command(self, tsn, command_id, args):
-        """Handle a ZDO command received on this cluster."""
-        pass
-
-
-class ZhaDeviceEntity(entity.Entity):
-    """A base class for ZHA devices."""
-
-    def __init__(self, device, manufacturer, model, application_listener,
-                 keepalive_interval=7200, **kwargs):
-        """Init ZHA endpoint entity."""
-        self._device_state_attributes = {
-            'nwk': '0x{0:04x}'.format(device.nwk),
-            'ieee': str(device.ieee),
-            'lqi': device.lqi,
-            'rssi': device.rssi,
-        }
-
-        ieee = device.ieee
-        ieeetail = ''.join(['%02x' % (o, ) for o in ieee[-4:]])
-        if manufacturer is not None and model is not None:
-            self._unique_id = "{}_{}_{}".format(
-                slugify(manufacturer),
-                slugify(model),
-                ieeetail,
-            )
-            self._device_state_attributes['friendly_name'] = "{} {}".format(
-                manufacturer,
-                model,
-            )
-        else:
-            self._unique_id = str(ieeetail)
-
-        self._device = device
-        self._state = 'offline'
-        self._keepalive_interval = keepalive_interval
-
-        application_listener.register_entity(ieee, self)
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID."""
-        return self._unique_id
-
-    @property
-    def state(self) -> str:
-        """Return the state of the entity."""
-        return self._state
-
-    @property
-    def device_state_attributes(self):
-        """Return device specific state attributes."""
-        update_time = None
-        if self._device.last_seen is not None and self._state == 'offline':
-            time_struct = time.localtime(self._device.last_seen)
-            update_time = time.strftime("%Y-%m-%dT%H:%M:%S", time_struct)
-            self._device_state_attributes['last_seen'] = update_time
-        if ('last_seen' in self._device_state_attributes and
-                self._state != 'offline'):
-            del self._device_state_attributes['last_seen']
-        self._device_state_attributes['lqi'] = self._device.lqi
-        self._device_state_attributes['rssi'] = self._device.rssi
-        return self._device_state_attributes
-
-    async def async_update(self):
-        """Handle polling."""
-        if self._device.last_seen is None:
-            self._state = 'offline'
-        else:
-            difference = time.time() - self._device.last_seen
-            if difference > self._keepalive_interval:
-                self._state = 'offline'
-            else:
-                self._state = 'online'
-
-
-def get_discovery_info(hass, discovery_info):
-    """Get the full discovery info for a device.
-
-    Some of the info that needs to be passed to platforms is not JSON
-    serializable, so it cannot be put in the discovery_info dictionary. This
-    component places that info we need to pass to the platform in hass.data,
-    and this function is a helper for platforms to retrieve the complete
-    discovery info.
-    """
-    if discovery_info is None:
-        return
-
-    discovery_key = discovery_info.get('discovery_key', None)
-    all_discovery_info = hass.data.get(DISCOVERY_KEY, {})
-    return all_discovery_info.get(discovery_key, None)
-
-
-async def safe_read(cluster, attributes, allow_cache=True, only_cache=False):
-    """Swallow all exceptions from network read.
-
-    If we throw during initialization, setup fails. Rather have an entity that
-    exists, but is in a maybe wrong state, than no entity. This method should
-    probably only be used during initialization.
-    """
-    try:
-        result, _ = await cluster.read_attributes(
-            attributes,
-            allow_cache=allow_cache,
-            only_cache=only_cache
-        )
-        return result
-    except Exception:  # pylint: disable=broad-except
-        return {}
-
-
-async def configure_reporting(entity_id, cluster, attr, skip_bind=False,
-                              min_report=300, max_report=900,
-                              reportable_change=1):
-    """Configure attribute reporting for a cluster.
-
-    while swallowing the DeliverError exceptions in case of unreachable
-    devices.
-    """
-    from zigpy.exceptions import DeliveryError
-
-    attr_name = cluster.attributes.get(attr, [attr])[0]
-    cluster_name = cluster.ep_attribute
-    if not skip_bind:
-        try:
-            res = await cluster.bind()
-            _LOGGER.debug(
-                "%s: bound  '%s' cluster: %s", entity_id, cluster_name, res[0]
-            )
-        except DeliveryError as ex:
-            _LOGGER.debug(
-                "%s: Failed to bind '%s' cluster: %s",
-                entity_id, cluster_name, str(ex)
-            )
-
-    try:
-        res = await cluster.configure_reporting(attr, min_report,
-                                                max_report, reportable_change)
-        _LOGGER.debug(
-            "%s: reporting '%s' attr on '%s' cluster: %d/%d/%d: Result: '%s'",
-            entity_id, attr_name, cluster_name, min_report, max_report,
-            reportable_change, res
-        )
-    except DeliveryError as ex:
-        _LOGGER.debug(
-            "%s: failed to set reporting for '%s' attr on '%s' cluster: %s",
-            entity_id, attr_name, cluster_name, str(ex)
         )
