+++ homeassistant/components/zha/entities/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from .entity import Entity
+from .device_entity import ZhaDeviceEntity
